Function json_io::load

pub fn load<T: Decodable>(path: &Path) -> Result<T, Error>

Construct a Decodable type from a JSON file at the given path. json_io will first try and open the file with the path exactly as given. If the file isn't found, it will set the extension to .json and try again.